backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "sobiamehak",email = "mehaksobi777@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "fastapi (>=0.68.0)",
    "uvicorn (>=0.15.0)",
    "qdrant-client (>=1.9.0)",
    "openai (>=1.0.0)",
    "tiktoken (>=0.4.0)",
    "pydantic (>=1.10.0)",
    "requests (>=2.28.0)",
    "python-multipart (>=0.0.6)",
    "passlib (>=1.7.4)",
    "python-jose (>=3.3.0)",
    "alembic (>=1.11.0)",
    "email-validator (>=2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
